requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-chroma>=0.1.0
langgraph>=0.1.0

# Data processing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Vector database
chromadb>=0.4.0

# Environment management
python-dotenv>=1.0.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Web UI
gradio>=4.0.0

# HTTP and SSL
urllib3>=2.0.0

# XML parsing
# removed: no pip package named 'xml'; use built-in xml module
